requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.4
anyio==3.3.3
apache-airflow==2.2.1
apache-airflow-providers-cncf-kubernetes==2.2.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
argcomplete==1.12.3
async-timeout==4.0.2
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
blinker==1.4
cachetools==4.2.4
cattrs==1.5.0
certifi==2020.12.5
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
croniter==1.0.15
cryptography==3.4.8
dateparser==1.1.0
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
distlib==0.3.4
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
filelock==3.4.0
Flask==1.1.4
Flask-AppBuilder==3.3.3
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
Flask-WTF==0.14.3
frozenlist==1.2.0
gevent==21.12.0
google==3.0.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-bigquery==2.31.0
google-cloud-core==2.2.1
google-cloud-datastore==2.4.0
google-cloud-language==2.3.1
google-cloud-pubsub==2.9.0
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
graphviz==0.17
greenlet==1.1.2
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httplib2==0.20.2
httpx==0.19.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
inflection==0.5.1
ipython==7.30.1
iso8601==0.1.16
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
joblib==1.1.0
jsonschema==3.2.0
kubernetes==11.0.0
lazy-object-proxy==1.4.3
libcst==0.3.23
lockfile==0.12.2
lxml==4.7.1
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
matplotlib-inline==0.1.3
multidict==5.2.0
mypy-extensions==0.4.3
nltk==3.6.6
numpy==1.21.5
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==21.0
pandas==1.3.5
parso==0.8.3
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prison==0.2.1
prompt-toolkit==3.0.24
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
py4j==0.10.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
pyspark==3.2.0
python-binance==1.0.15
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytrends==4.7.3
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==5.4.1
redis==4.0.2
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==10.12.0
rsa==4.8
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tqdm==4.62.3
traitlets==5.1.1
tweepy==3.10.0
typing-extensions==4.0.1
typing-inspect==0.7.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
wcwidth==0.2.5
websocket-client==1.2.3
websockets==9.1
Werkzeug==1.0.1
wrapt==1.13.3
WSGIserver==1.3
WTForms==2.3.3
yarl==1.7.2
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
